test: Remove some too-specific doc tests (for now)
It's a bit too early for tests which are this specific, which is making
it difficult to refactor the existing code structures. It's better to
keep the AST etc. flexible, and instead rely on higher-level functional
tests. More specific tests can be added only once the structures start
to solidify.