interpreter: Fix function calls with mismatched args and params
Allow function calls when a different number of arguments are supplied
compared to the function's declared parameters.

If there are more arguments than parameters then the extra arguments are
still evaluated, but are not passed to the function. If there are fewer
arguments than parameters then `undefined` is passed in place of the
missing arguments.