interpreter,runtime: Well-known identifier optimisation
Store hard-coded "well-known" identifiers as an enum for performance.

The advantage to this is that it's much faster to compare and hash
well-known identifiers, making it faster to lookup properties of the
global object, for example. The downside is that all identifiers are now
a bit larger in memory (24 -> 32 bytes).

This makes some programs much faster. As an example, the
`project_euler/p007_10001st_prime.js` test now takes ~942ms, down from
~7,710ms (release mode with slow bit uncommented). This test benefits so
much because it calls `Math.sqrt()` in a tight loop, and since `Math`
and `sqrt` are now well-known identifiers, `IdentifierReference` and
`Member` expressions for them are now _much_ more efficient to evaluate.

Eventually _all_ identifier names should be interned, but I still need
to decide how that should work when using the same interpreter for
multiple scripts (e.g. for the REPL).