test262: Add an extremelly basic test262 runner
Many of the test262 test cases currently cause the interpreter to crash.
First goal is to eliminate these crashes (or teach the test runner to
tolerate them) so that this can actually be useful.

At that point, the number of passing vs. failing test cases can start to
be tracked and graphed, and then iteratively improved with Graph Driven
Development (GDD).