ast,parser,interpreter: Remove the idea of a `MemberExpression`
My hope was to split out a specific kind of `Expression` for the
expressions which are legal as the left-hand-side of, for example

- member access operator
- function call operator
- assignment operator
- etc.

However, after thinking about it this doesn't really make sense. There
are very few expressions which _aren't_ legal in those positions and
it's non-trivial to work out which ones! For example:

The parser should allow this:

```jakescript
("hello" + 2).length
```

The parser should not allow this:

```jakescript
2.length
```

But should the parser allow this??:

```jakescript
(2 + 3).length
```

If the parser is made to do anything to validate this then it will need
to happen outside of the type system, unfortunately. Failing that, it
will need to be done at runtime once the expression has been evaluated
to a value.